% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{losses}
\alias{losses}
\title{A combination of all the losses in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `losses` data frame combines all six lines of business into one data frame
and removes the premium and meta information columns that do not need to
be recorded for every development period.}
\format{A data frame with 77900 observations of 8 variables
\describe{
\item{lob}{line of business}
\item{gr_code}{NAIC company code (includes insurer groups and single insurers)}
\item{gr_name}{NAIC company name (includes insurer groups and single insurers)}
\item{origin}{year in which accidents occurred}
\item{dev}{Years of development}
\item{incurred}{Incurred losses and allocated expenses at year end}
\item{paid}{cumulative paid losses and defense and cost containment expenses at year end}
\item{bulk}{Bulk and IBNR reserves on net losses and defense and cost containment expenses reported
at year end}
}}
\usage{
losses
}
\description{
A combination of all the losses in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `losses` data frame combines all six lines of business into one data frame
and removes the premium and meta information columns that do not need to
be recorded for every development period.
}
\examples{
losses
}
\keyword{datasets}

