% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{othliab}
\alias{othliab}
\title{Other liability (occurrence basis) loss reserve data}
\format{A data frame with 23900 observations and 13 variables:
\describe{
\item{GRCODE}{NAIC company code (includes insurer groups and single insurers)}
\item{GRNAME}{NAIC company name (includes insurer groups and single insurers)}
\item{AccidentYear}{year in which accidents occurred}
\item{DevelopmentYear}{year of evaluation}
\item{DevelopmentLag}{Years of development (AccidentYear - DevelopmentYear + 1)}
\item{IncurLoss_h1}{Incurred losses and allocated expenses at year end}
\item{CumPaidLoss_h1}{cumulative paid losses and defense and cost containment expenses at year end}
\item{BulkLoss_h1}{Bulk and IBNR reserves on net losses and defense and cost containment expenses reported
at year end}
\item{EarnedPremDIR_h1}{direct and assumed premiums earned at incurral year}
\item{EarnedPremCeded_h1}{ceded premiums earned at incurral year}
\item{EarnedPremNet_h1}{net premiums earned at incurral year}
\item{Single}{factor for determining if entity is a single insurer (1) or a group insurer (0)}
\item{PostedReserve97_h1}{posted reserves in year 1997 taken from the Underwriting and Investment Exhibit -
Part 2A, including net losses unpaid and unpaid loss adjustment expense}
}}
\usage{
othliab
}
\description{
Data from the NAIC schedule P.  Data collected and cleaned by
Glenn G. Meyers and Peng Shi. The data set is freely available
at \url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
}
\examples{
othliab
}
\keyword{datasets}

