% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium}
\alias{premium}
\title{A combination of all the premium in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `premium` data frame combines all six lines of business into one data frame
and removes the losses and meta information columns that are not recorded on a
origin year basis.}
\format{A data frame with 7790 observations of 7 variables
\describe{
\item{lob}{Line of Business}
\item{gr_code}{NAIC company code (includes insurer groups and single insurers)}
\item{gr_name}{NAIC company name (includes insurer groups and single insurers)}
\item{origin}{year in which accidents occurred}
\item{premium_earned}{direct and assumed premiums earned at incurral year}
\item{premium_ceded}{ceded premiums earned at incurral year}
\item{premium_net}{net premiums earned at incurral year}
}}
\usage{
premium
}
\description{
A combination of all the premium in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `premium` data frame combines all six lines of business into one data frame
and removes the losses and meta information columns that are not recorded on a
origin year basis.
}
\examples{
premium
}
\keyword{datasets}

