% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reserves}
\alias{reserves}
\title{A combination of all the premium in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `reserves` data frame combines all six lines of business into one data frame
and removes the losses and premium information columns that are not recorded on a
company basis.}
\format{A data frame with 7790 observations of 7 variables
\describe{
\item{lob}{Line of Business}
\item{gr_code}{NAIC company code (includes insurer groups and single insurers)}
\item{gr_name}{NAIC company name (includes insurer groups and single insurers)}
\item{single}{factor for determining if entity is a single insurer (1) or a group insurer (0)}
\item{reserve_1997}{posted reserves in year 1997 taken from the Underwriting and Investment Exhibit -
Part 2A, including net losses unpaid and unpaid loss adjustment expense}
}}
\usage{
reserves
}
\description{
A combination of all the premium in the data sets on the case website.
\url{http://www.casact.org/research/index.cfm?fa=loss_reserves_data}
the `reserves` data frame combines all six lines of business into one data frame
and removes the losses and premium information columns that are not recorded on a
company basis.
}
\examples{
reserves
}
\keyword{datasets}

